+++ nta.utils/nta/utils/test_utils/amqp_test_utils.py
@@ -33,10 +33,10 @@ import uuid
 from mock import patch
 import requests
 
-import nta.utils.amqp
-from nta.utils.amqp import (AMQPErrorCodes,
-                            getRabbitmqConnectionParameters,
-                            RabbitmqManagementConnectionParams)
+from nta.utils.amqp import connection as amqp_connection
+from nta.utils.amqp import constants as amqp_constants
+from nta.utils.amqp import exceptions as amqp_exceptions
+from nta.utils.amqp import synchronous_amqp_client
 from nta.utils.test_utils.config_test_utils import ConfigAttributePatch
 
 
@@ -143,7 +143,7 @@ class RabbitmqVirtualHostPatch(object):
     assert not self.active
 
     # Use RabbitMQ Management Plugin to create the new temporary vhost
-    connectionParams = RabbitmqManagementConnectionParams()
+    connectionParams = amqp_connection.RabbitmqManagementConnectionParams()
 
     url = "http://%s:%s/api/vhosts/%s" % (
       connectionParams.host, connectionParams.port, self._vhost)
@@ -195,7 +195,7 @@ class RabbitmqVirtualHostPatch(object):
 
       # Apply a config patch to override the rabbitmq virtual host to be
       # used by message_bus_connector and others
-      rabbitmqConfig = nta.utils.amqp.RabbitmqConfig()
+      rabbitmqConfig = amqp_connection.RabbitmqConfig()
       self._configPatch = ConfigAttributePatch(
         rabbitmqConfig.CONFIG_NAME,
         rabbitmqConfig.baseConfigDir,
@@ -207,7 +207,8 @@ class RabbitmqVirtualHostPatch(object):
                         self.__class__.__name__, self._vhost)
 
       # Self-validation
-      connectionParams = getRabbitmqConnectionParameters()
+      connectionParams = (
+        amqp_connection.getRabbitmqConnectionParameters())
       actualVhost = connectionParams.vhost
       assert actualVhost == self._vhost, (
         "Expected vhost=%r, but got vhost=%r") % (self._vhost, actualVhost)
@@ -245,7 +246,8 @@ class RabbitmqVirtualHostPatch(object):
   def _deleteTemporaryVhost(self):
     """ Delete a RabbitMQ virtual host """
     # Use RabbitMQ Management Plugin to delete the temporary vhost
-    connectionParams = RabbitmqManagementConnectionParams()
+    connectionParams = (
+        amqp_connection.RabbitmqManagementConnectionParams())
 
     url = "http://%s:%s/api/vhosts/%s" % (
       connectionParams.host, connectionParams.port, self._vhost)
@@ -286,16 +288,17 @@ def managedExchangeDeleter(exchange):
     else:
       messageExchanges = exchange
 
-    connParams = getRabbitmqConnectionParameters()
+    connParams = amqp_connection.getRabbitmqConnectionParameters()
 
     for exchangeName in messageExchanges:
       try:
-        with nta.utils.amqp.SynchronousAmqpClient(connParams) as amqpClient:
+        with synchronous_amqp_client.SynchronousAmqpClient(connParams) as (
+          amqpClient):
           _LOGGER.info("managedExchangeDeleter: Deleting exchange=%r",
                        exchangeName)
           amqpClient.deleteExchange(exchangeName, ifUnused=False)
-      except nta.utils.amqp.AmqpChannelError as e:
-        if e.code == AMQPErrorCodes.NOT_FOUND:
+      except amqp_exceptions.AmqpChannelError as e:
+        if e.code == amqp_constants.AMQPErrorCodes.NOT_FOUND:
           _LOGGER.info("managedExchangeDeleter: exchange=%r not found (%r)",
                        exchangeName, e)
         else:
@@ -320,17 +323,18 @@ def managedQueueDeleter(mq):
     else:
       messageQueues = mq
 
-    connParams = getRabbitmqConnectionParameters()
+    connParams = amqp_connection.getRabbitmqConnectionParameters()
 
     for mqName in messageQueues:
       try:
-        with nta.utils.amqp.SynchronousAmqpClient(connParams) as amqpClient:
+        with synchronous_amqp_client.SynchronousAmqpClient(connParams) as (
+          amqpClient):
           _LOGGER.info("managedQueueDeleter: Deleting mq=%r",
                        mqName)
           amqpClient.deleteQueue(mqName, ifUnused=False, ifEmpty=False)
-      except nta.utils.amqp.AmqpChannelError as e:
+      except amqp_exceptions.AmqpChannelError as e:
         # Suppress queue_delete error (perhaps queue was never created?)
-        if e.code == AMQPErrorCodes.NOT_FOUND:
+        if e.code == amqp_constants.AMQPErrorCodes.NOT_FOUND:
           _LOGGER.info("managedQueueDeleter: mq=%r not found (%r)",
                        mqName, e)
         else:
